Remove FF_GITHUB_ETAG_CACHE flag; fold ETag cache into GithubAPIClient
The project has no feature-flag mechanism, so drop the FF_GITHUB_ETAG_CACHE
kill-switch and collapse the two-client split. The ETag/304 response cache
(memoization + If-None-Match revalidation, best-effort) now lives directly in
GithubAPIClient and is always on; the CachedGithubAPIClient decorator and the
github_api_client.raw indirection are removed.

Behavior is unchanged from production (the flag was already 1 there), so the
304 rate-limit savings are preserved. The cache repository, table, migrations
and purge CLI are untouched. Cache test coverage moved into GithubAPIClientTest.
