+++ .env
@@ -33,7 +33,6 @@ SLACK_CLIENT_SECRET=slack_client_secret
 GITHUB_APP_HOMEPAGE_URL=https://github.com/apps/slub-yeee
 CHECKS_BLACKLIST=
 FF_RECORD_EVENTS=0
-FF_GITHUB_ETAG_CACHE=1
 ###< Slub Application ###
 
 ###> sentry/sentry-symfony ###
