+++ .env.test
@@ -18,7 +18,6 @@ SLACK_CLIENT_SECRET=slack_client_secret
 GITHUB_APP_HOMEPAGE_URL=https://github.com/apps/slub-yeee-test
 CHECKS_BLACKLIST=check1,check2
 FF_RECORD_EVENTS=1
-FF_GITHUB_ETAG_CACHE=1
 ###< Slub Application ###
 
 ###> sentry/sentry-symfony ###
